Move package of test cases
